+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- イベント検索からX-Total-Countを削除
+- 販売者検索からX-Total-Countを削除
+
 ### Deprecated
 
 ### Removed